+++ test/test_various.jl
@@ -208,8 +208,8 @@ end
 
 @testset "Test simulation pretty printing" begin
     sim = mr.Simulation([mr.dwi(bval=1, TR=2000), mr.dwi(bval=2)], geometry=mr.Spheres(radius=[1, 2.], repeats=[5, 5, 5]), R1=0.1, surface_relaxivity=0.3)
-    @test repr(sim, context=:compact => true) == "Simulation(2 sequences, Geometry(2 repeating Round objects, ), D=0.0um^2/ms, GlobalProperties(R1=0.1kHz, ))" 
-    @test repr(sim, context=:compact => false) == "Simulation(Geometry(2 repeating Round objects, ), D=0.0um^2/ms, GlobalProperties(R1=0.1kHz, )):
+    @test repr(sim, context=:compact => true) == "Simulation(2 sequences, Geometry(2 repeating Round objects, ), D=3.0um^2/ms, GlobalProperties(R1=0.1kHz, ))" 
+    @test repr(sim, context=:compact => false) == "Simulation(Geometry(2 repeating Round objects, ), D=3.0um^2/ms, GlobalProperties(R1=0.1kHz, )):
 2 sequences:
 Sequence (TR=2000.0ms):
     - InstantRFPulse: t=0.0ms, θ=90.0°, ϕ=-90.0°;